Remove block_info.go
